+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 ### Added
 - Check tye version at startup
 - Action to update tye global tool
-- Action to perform shutdown tye host
+- Action to perform shutdown tye application
+- Action to connect to tye host
 ### Changed
 - Plugin icon
 - Update tye version to 0.7.0-alpha.21279.2